work around another bug
here's weird one
* WARNING: Method invocation on a nil value detected.
The method < was invoked on reset_desired? within args.state, but
reset_desired? has a
value of nil (the invocation of < was supressed). Consider setting
a default value for reset_desired? to avoid null reference exceptions
later on."
 but there are only 3 instances of reset_desired?
@gtk_state[:reset_desired?] = false
 in my class init,
     if @kb.key_down.r then
       @gtk_state[:reset_desired?] = true
     end
 in my input handler, and
$gtk.reset seed: rand(Time.now.sec) if args.state[:reset_desired?]
 in my tick.  The warning happens when I'm spamming keys other than "r".
If I replace all of that and move the $gtk.reset into the if
@kb.key_down.r, then no sporadic warnings.